+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-fewsnorm
 0.14 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Work towards fix for sync_track_record_cache method of FewsNormSource.
 
 
 0.13 (2012-04-12)